requirements.txt
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.11.0
environ==1.0
freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
